pyproject.toml
[tool.poetry]
name = "yt-music-dl"
version = "0.1.0"
description = "Downloads songs from youtube, renames them and writes the song metadata"
license = "GPL-3.0-only"
authors = ["Juho Röyskö <roysko.juho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
invoke = "^1.7.0"
pydub = "^0.25.1"
pygame = "^2.1.2"
yt-dlp = "^2022.4.8"
pathvalidate = "^2.5.0"
requests = "^2.28.1"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-dotenv = "^0.5.2"
coverage = "^6.3.2"
pylint = "^2.13.5"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']